move astype tests to test_dtypes
